+++ spec/rubocop/cop/lint/shadowed_exception_spec.rb
@@ -60,6 +60,20 @@ describe RuboCop::Cop::Lint::ShadowedException do
       RUBY
     end
 
+    it 'accepts rescuing multiple same error code exceptions' do
+      # System dependent error code depends on runtime environment.
+      stub_const('Errno::EAGAIN::Errno', 35)
+      stub_const('Errno::EWOULDBLOCK::Errno', 35)
+
+      expect_no_offenses(<<-RUBY.strip_indent)
+        begin
+          something
+        rescue Errno::EAGAIN, Errno::EWOULDBLOCK
+          handle_exception
+        end
+      RUBY
+    end
+
     it 'registers an offense rescuing exceptions that are ' \
       'ancestors of each other ' do
       inspect_source(<<-RUBY.strip_indent)