+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * [#4830](https://github.com/bbatsov/rubocop/issues/4830): Prevent `Lint/BooleanSymbol` from truncating symbol's value in the message when offense is located in the new syntax hash. ([@akhramov][])
 * [#4747](https://github.com/bbatsov/rubocop/issues/4747): Fix `Rails/HasManyOrHasOneDependent` cop incorrectly flags `with_options` blocks. ([@koic][])
 * [#4836](https://github.com/bbatsov/rubocop/issues/4836): Make `Rails/OutputSafety` aware of safe navigation operator. ([@drenmi][])
+* [#4843](https://github.com/bbatsov/rubocop/issues/4843): Fix `Lint/ShadowedException` cop aware of same system error code. ([@koic][])
 
 ### Changes